Improved jcf.range to calculate handles' position when multiple handlers overlap
By default, multiple handles will not overlap by a ratio of 0.55 which
matches the current demos. A new option  has been
added. It expects values between 0 and 2, allowing the user to set the
desired offset to which handlers will touch each other without overla-
pping, based on the handle's width as a base size for the ratio calc.